+++ www/index.css
@@ -1,0 +1,27 @@
+body {
+  font-family: sans-serif;
+}
+
+[data-focus] {
+  display: flex;
+  flex-direction: row;
+  border: 1px solid blue;
+  padding: 5px;
+  margin: 10px;
+  background-color: rgba(0, 0, 255, 0.2);
+}
+
+[data-focus][data-orientation='vertical'] {
+  flex-direction: column;
+  border: 1px solid red;
+  background-color: rgba(255, 0, 0, 0.2);
+}
+
+:focus {
+  outline: 0;
+}
+
+:focus,
+.focused {
+  background-color: yellow;
+}
